% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{remove_var_na}
\alias{remove_var_na}
\title{Remove Columns with too many NA values}
\usage{
remove_var_na(df, d.count.na, tolerance = 75, verbose = TRUE)
}
\description{
This function Remove Columns with too many NA values
}
\details{
details
}
\examples{
\dontrun{
remove_var_na(df, d.count.na, tolerance = 0.3, verbose = TRUE)
}
}
