% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{to_factor}
\alias{to_factor}
\title{Convert variables to factor}
\usage{
to_factor(df, vf)
}
\arguments{
\item{df}{dataframe}

\item{d.count}{dataframe from na_count function}

\item{tolerance}{percentual maximum of NA tolerable}

\item{verbose}{verbose}
}
\description{
This function  Convert variables to factor
}
\details{
Convert variables to factor
}
